backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
gradio_client
requests
imagekitio
google-genai
pillow